+++ src/manager.rs
@@ -2,36 +2,8 @@ use std::future::Future;
 
 use crate::address::Address;
 use crate::context::Context;
-use crate::envelope::{BroadcastMessageEnvelope, MessageEnvelope};
 use crate::spawn::Spawner;
 use crate::Actor;
-
-/// A message that can be sent by an Address to the manage loop
-pub(crate) enum AddressMessage<A> {
-    /// A message from the last address telling the actor that it should shut down
-    LastAddress,
-    /// A message being sent to the actor. To read about envelopes and why we use them, check out
-    /// `envelope.rs`
-    Message(Box<dyn MessageEnvelope<Actor = A>>),
-}
-
-/// A message that can be sent by another actor on the same address to the manage loop
-pub(crate) enum BroadcastMessage<A> {
-    /// A message from another actor on the same address telling the actor to unconditionally shut
-    /// down.
-    Shutdown,
-    Message(Box<dyn BroadcastMessageEnvelope<Actor = A>>),
-}
-
-impl<A> Clone for BroadcastMessage<A> {
-    fn clone(&self) -> Self {
-        use self::BroadcastMessage::*;
-        match self {
-            Shutdown => Shutdown,
-            Message(msg) => Message(msg.clone()),
-        }
-    }
-}
 
 /// If and how to continue the manage loop
 #[derive(PartialEq, Eq, Debug, Copy, Clone)]
