+++ src/message_channel.rs
@@ -6,15 +6,14 @@ use std::fmt::Debug;
 
 use futures_core::future::BoxFuture;
 use futures_core::stream::BoxStream;
-use futures_sink::Sink;
 
 use crate::address::{Address, WeakAddress};
+use crate::envelope::ReturningEnvelope;
 use crate::private::Sealed;
 use crate::receiver::Receiver;
 use crate::refcount::{RefCounter, Shared, Strong};
 use crate::send_future::{ResolveToHandlerReturn, SendFuture};
 use crate::{Handler, KeepRunning};
-use crate::envelope::ReturningEnvelope;
 
 /// A message channel is a channel through which you can send only one kind of message, but to
 /// any actor that can handle it. It is like [`Address`](../address/struct.Address.html), but associated with
